pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "git-pulse"
version = "0.1.0"
description = "Background Git repository updater — keeps your default branches fresh while you work"
readme = "README.md"
license = {text = "MIT"}
requires-python = ">=3.10"
authors = [
    {name = "Vedanth Vasudev"},
]
keywords = ["git", "automation", "cli", "daemon", "sync"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Environment :: Console",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Operating System :: MacOS",
    "Operating System :: POSIX :: Linux",
    "Programming Language :: Python :: 3",
    "Topic :: Software Development :: Version Control :: Git",
]
dependencies = [
    "typer>=0.9.0",
    "pyyaml>=6.0",
    "gitpython>=3.1.40",
    "rich>=13.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0",
    "pytest-cov>=4.0",
]

[project.scripts]
git-pulse = "git_pulse.cli:app"

[tool.setuptools.packages.find]
include = ["git_pulse*"]

[tool.pytest.ini_options]
testpaths = ["tests"]
markers = [
    "unit: Pure logic tests, no git repos or network (deselect with '-m not unit')",
    "integration: Tests with real git repos on disk (deselect with '-m not integration')",
    "e2e: End-to-end CLI tests via CliRunner (deselect with '-m not e2e')",
]

[tool.coverage.run]
source = ["git_pulse"]

[tool.coverage.report]
show_missing = true
exclude_lines = [
    "pragma: no cover",
    "if __name__",
    "raise typer.Exit",
]
